+++ proof_ics23.go
@@ -30,10 +30,7 @@ func (t *ImmutableTree) VerifyMembership(proof *ics23.CommitmentProof, key []byt
 	if err != nil {
 		return false, err
 	}
-	root, err := t.Hash()
-	if err != nil {
-		return false, err
-	}
+	root := t.Hash()
 
 	return ics23.VerifyMembership(ics23.IavlSpec, root, proof, key, val), nil
 }
@@ -93,10 +90,7 @@ func (t *ImmutableTree) GetNonMembershipProof(key []byte) (*ics23.CommitmentProo
 
 // VerifyNonMembership returns true iff proof is a NonExistenceProof for the given key.
 func (t *ImmutableTree) VerifyNonMembership(proof *ics23.CommitmentProof, key []byte) (bool, error) {
-	root, err := t.Hash()
-	if err != nil {
-		return false, err
-	}
+	root := t.Hash()
 
 	return ics23.VerifyNonMembership(ics23.IavlSpec, root, proof, key), nil
 }
@@ -104,10 +98,7 @@ func (t *ImmutableTree) VerifyNonMembership(proof *ics23.CommitmentProof, key []
 // createExistenceProof will get the proof from the tree and convert the proof into a valid
 // existence proof, if that's what it is.
 func (t *ImmutableTree) createExistenceProof(key []byte) (*ics23.ExistenceProof, error) {
-	_, err := t.Hash()
-	if err != nil {
-		return nil, err
-	}
+	t.Hash()
 	path, node, err := t.root.PathToLeaf(t, key, t.version+1)
 	nodeVersion := t.version + 1
 	if node.nodeKey != nil {