+++ tree_fuzz_test.go
@@ -113,10 +113,9 @@ func TestMutableTreeFuzz(t *testing.T) {
 
 	for size := 5; iterations < maxIterations; size++ {
 		for i := 0; i < progsPerIteration/size; i++ {
-			tree, err := getTestTree(0)
-			require.NoError(t, err)
+			tree := getTestTree(0)
 			program := genRandomProgram(size)
-			err = program.Execute(tree)
+			err := program.Execute(tree)
 			if err != nil {
 				str, err := tree.String()
 				require.Nil(t, err)